refactor(#1375): derive types/* unions from enums.ts SSoT
Inline literal unions in types/{vehicle,search-result,customer,vehicle-detail}
were re-spellings of enums.ts closed sets, so renaming an enum member compiled
green while shipping a runtime mismatch. Derive each from the SSoT instead:

- vehicle.ts: import + re-export Transmission/VehicleStatus from enums
- search-result.ts: SearchResultKind = BookingFulfillmentMode; transmission: Transmission
- customer.ts: status: BookingStatus
- vehicle-detail.ts: source: BookingSource; status: Extract<BookingStatus,'CONFIRMED'|'ACTIVE'>
- fold REVIEW_DIMENSIONS into enums.ts beside its sibling review enums; re-export from review.ts

Type-only, no migration; all three packages typecheck unchanged.

Closes #1375
Part of #1369
